src/service.c: don't try to show any lockscreen shortcut on Lomiri
Lomiri itself doesn't read any schema for the lockscreen [1]. It also
won't show the "accel" or shortcut property of a GMenuItem [2]. Thus, I
think it's not worth it to add a schema in lomiri-schemas just to be
ignored by Lomiri both ways. As such, remove the functionality to read
a shortcut for lockscreen on Lomiri.

While we're at it, also make sure that `p->keybinding_settings` is
initialized even if it's on non of the recognized DE.

[1] https://gitlab.com/ubports/development/core/lomiri/-/blob/main/qml/Components/Dialogs.qml#L121-124
[2] https://gitlab.com/ubports/development/core/lomiri/-/blob/main/qml/Panel/Indicators/IndicatorMenuItemFactory.qml